+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 Version 0.7.2 - (2010-Sep-06)
 ==============================
+ * Added ``form_template`` in :class:`djpcms.views.appsite.ModelApplication`` for customizing uniform rendering.
  * updated to jQuery 1.4.2 and added two jQuery UI themes.
  * Added :setting:`DJPCMS_STYLE` setting for specifying ``css`` style.
  * :class:`djpcms.views.cache.PageCache` works when django sessions are not available. Previously it was failing.